pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clinical-trial-eligibility-screener"
version = "0.1.0"
description = "OpenEnv clinical trial eligibility screener environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=2.30.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.6.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
py-modules = ["baseline", "client", "grader", "inference", "models"]

[tool.setuptools.package-data]
server = ["data/*.json"]

[tool.ruff]
line-length = 100
target-version = "py310"
